*** Project completely re-arranged
* Code Changes
- Upgraded to Swift 4.
- Added storyboard and .swift file for table.
- Removed Tests and its scheme since indicator cannot be tested via unit testing.
- Added DemoUITests and its scheme.
- LaunchScreen.storyboard updated.

* Other Changes
- travis.yml file upgraded for DemoUITests.
- Project version is changed to 2.0.0
- Podspec file upgraded.

* TODO's
- Project owner should push to CocoaPods with tag 2.0.0
