+++ .travis.yml
@@ -2,5 +2,5 @@ osx_image: xcode7.1
 language: objective-c
 cache: cocoapods
 before_install: gem install xcpretty cocoapods obcd slather -N
-script: xcodebuild -project Demo.xcodeproj -scheme Tests -sdk iphonesimulator build test GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES clean test | xcpretty -c && exit ${PIPESTATUS[0]}
+script: xcodebuild -project Demo.xcodeproj -scheme DemoUITests -sdk iphonesimulator build test GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES clean test | xcpretty -c && exit ${PIPESTATUS[0]}
 after_success: slather
