+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 Check https://github.com/3lvis/NetworkActivityIndicator/releases for more information.
+
+# V. 2.0.0
+
+- Upgraded to Swift 4
+- Added UITests
